添加异常信息；增加下载错误状态

+++ DownloaderLibrary/src/main/java/com/excellence/downloader/utils/MemorySpaceCheck.java
@@ -42,7 +42,7 @@ public class MemorySpaceCheck {
 			return getSystemAvailableSize() > length;
 		}
 	}
-	public static boolean hasSDEnoughMemory(String filePath, int length) {
+	public static boolean hasSDEnoughMemory(String filePath, long length) {
 		Log.d("hasSD", getAvailableSize(filePath) + "::" + length);
 		return getAvailableSize(filePath) > length;
 	}
